chore: reducing Docker image layers (38 -> 7)
Only the instructions RUN, COPY, and ADD create layers. Other
instructions create temporary intermediate images, and do not
increase the size of the build (number of layers). However,
there is a limit of 127 layers in the final image. It is
therefore recommended to offer Dockerfiles in which the number
of layers created has been reduced.

Always combine RUN apt-get update with apt-get install in the
same RUN statement is a common way to do so. For details, see:

- https://emmer.dev/blog/reducing-docker-layers
- https://docs.docker.com/build/building/best-practices/#run

Test with:

  docker inspect tiacsys/tsn-asdf-ws:local | \
         jq '.[0].RootFS.Layers[]' | wc -l

Possible to squash:

  docker image save tiacsys/tsn-asdf-ws:local | \
  docker import - tiacsys/tsn-asdf-ws:squash && \
  docker inspect tiacsys/tsn-asdf-ws:squash | \
         jq '.[0].RootFS.Layers[]' | wc -l

Signed-off-by: Stephan Linz <linz@li-pro.net>